refactor(admin): address review feedback on artists pagination
Drop the unused useArtistsPageQuery wrapper and useArtistSorting hook
(inlined as a plain toggle in the route), keep previous page data
while a new page loads, and prefetch the loader using the actual URL
search params via loaderDeps instead of hardcoded defaults.
