+++ src/pages/admin/ArtistsManagement/types.ts
@@ -1,0 +1,8 @@
+import type { AdminArtistsSortKey } from "@/lib/searchSchemas";
+
+export type SortingKey = AdminArtistsSortKey;
+
+export type SortConfig = {
+  key: SortingKey;
+  direction: "asc" | "desc";
+};
